% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spotfuzz.r
\name{fuzzy_search_spotify}
\alias{fuzzy_search_spotify}
\title{Search Spotify for a artist/song combination.}
\usage{
fuzzy_search_spotify(artists, tracks, progress = FALSE)
}
\arguments{
\item{artists}{string with artist name}

\item{tracks}{string with song name}

\item{progress}{logical. If true, show track being searched and whether search is successful}
}
\value{
tibble(artist, track name, album name, track number, release date and spotify track uri)
}
\description{
Return essential metadata as a tibble containing
artist, track name, album name, track number, release date and spotify track uri.
This will try to iteratively loosen the search parameters if no result is returned.
First, it will see if the artist can be stripped to a single name.
Then it will try using just the longest word in the song title and finally
the second longest.  Will return a tibble with all NAs if search fails.
False positives with song words in common are rejected if the whole title is
substially different.
Vectorized along artist and track name.
It's not perfect.
}
\examples{
fuzzy_search_spotify("Rush","Tom Sawyer")
}
