% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzy_spotify_search.R, R/spotfuzz.r
\name{quick_search_spotify}
\alias{quick_search_spotify}
\title{Search Spotify for a artist/song combination.}
\usage{
quick_search_spotify(artist, track)

quick_search_spotify(artist, track)
}
\arguments{
\item{artist}{string}

\item{track}{string}
}
\value{
tibble(artist, track name, album name, track number, release date and spotify track uri)

tibble(artist, track name, album name, track number, release date and spotify track uri)
}
\description{
Return essential metadata as a tibble containing
artist, track name, album name, track number, release date and spotify track uri.

Return essential metadata as a tibble containing
artist, track name, album name, track number, release date and spotify track uri.
}
\examples{
quick_search_spotify("Rush","Tom Sawyer")
quick_search_spotify("Rush","Tom Sawyer")
}
