% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spotfuzz-package.R
\docType{package}
\name{spotfuzz-package}
\alias{spotfuzz}
\alias{spotfuzz-package}
\title{spotfuzz: Crude fuzzy search for spotifyr package}
\description{
Crude fuzzy search for spotifyr package.  Iteratively changes search string to get looser if item is not found initially.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/apsteinmetz/spotfuzz}
  \item Report bugs at \url{https://github.com/apsteinmetz/spotfuzz/issues}
}

}
\author{
\strong{Maintainer}: Art Steinmetz \email{apsteinmetz@yahoo.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
