% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spotfuzz.r
\name{quick_search_spotify}
\alias{quick_search_spotify}
\title{Search Spotify for a artist/song combination.}
\usage{
quick_search_spotify(artist, track)
}
\arguments{
\item{artist}{string with artist name}

\item{track}{string with song name}
}
\value{
tibble(artist, track name, album name, track number, release date and spotify track uri)
}
\description{
Return essential metadata as a tibble containing
artist, track name, album name, track number, release date and spotify track uri.
}
\examples{
quick_search_spotify("Rush","Tom Sawyer")
}
